requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.0.3
alembic==1.7.7
altair==4.2.0
anyio==3.5.0
arff==0.9
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.4.0
autopage==0.5.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==21.12b0
bleach==4.1.0
blinker==1.4
boto3==1.20.21
botocore==1.23.21
bravado==11.0.3
bravado-core==5.17.0
brotlipy==0.7.0
cachetools==5.0.0
captum>=0.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cliff==3.10.1
cloudpickle==2.0.0
cmaes==0.8.2
cmake==3.26.3
cmd2==2.4.1
colorama==0.4.4
colorlog==6.6.0
configparser==5.2.0
contourpy==1.0.7
cryptography==36.0.1
cycler==0.11.0
cytoolz==0.11.2
dask==2022.2.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dice-ml==0.7.2
docker-pycreds==0.4.0
efficientnet-pytorch>=0.7.1
entrypoints==0.3
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
Flask==2.1.3
flatbuffers==2.0
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.17.2
google-auth-oauthlib==0.4.6
grad-cam>=1.4.6
greenlet==1.1.2
grpcio==1.53.0
h5py==3.6.0
idna==3.3
imagecodecs==2022.2.22
imageio==2.16.1
imbalanced-learn==0.9.1
imblearn==0.0
imgaug==0.4.0
importlib-metadata==4.11.2
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonpointer==2.2
jsonref==0.2
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-events==0.6.3
jupyter-server==1.23.6
jupyter_core==5.3.0
jupyter_server_terminals==0.4.4
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
jupyterlab-widgets==1.1.0
kiwisolver==1.3.2
krippendorff==0.5.1
lazy_loader==0.2
lightning-utilities==0.8.0
lit==16.0.1
locket==0.2.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==2.0.5
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.2
munch==2.5.0
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.5.10
nbconvert==7.3.1
nbformat==5.2.0
neptune-client==0.13.3
nest-asyncio==1.5.4
networkx==2.7
notebook==6.4.8
notebook-shim==0.1.0
numpy==1.22.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
onnxruntime==1.10.0
opencv-python==4.5.5.62
opencv-python-headless==4.7.0.72
optuna==2.10.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
pathtools==0.1.2
patsy==0.5.2
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.2.0
plotly==5.13.1
pretrainedmodels>=0.7.4
prettytable==3.2.0
prometheus-client==0.13.0
promise==2.3
prompt-toolkit==3.0.26
protobuf==3.20.1
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydeck==0.7.1
pyDeprecate==0.3.1
Pygments==2.11.2
PyJWT==2.3.0
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-fid>=0.3.0
pytorch-gradcam==0.2.1
pytorch-lightning>=2.0.1.post0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyu2f==0.1.5
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rsa==4.8
s3transfer==0.5.0
scikit-image==0.19.1
scikit-learn==1.1.0
scikit-multilearn==0.2.0
scipy==1.8.0
seaborn==0.11.2
segmentation-models-pytorch==0.3.0
semver==2.13.0
Send2Trash==1.8.0
sentry-sdk==1.5.0
Shapely==1.8.0
shortuuid==1.0.8
SimpleITK==2.2.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.4
SQLAlchemy==1.4.35
stack-data==0.1.4
statsmodels==0.13.1
stevedore==3.5.0
streamlit==1.8.1
strict-rfc3339==0.7
subprocess32==3.5.4
swagger-spec-validator==2.7.4
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2022.2.9
tim==0.5
timm>=0.4.12
tinycss2==1.2.1
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
torch>=2.0.0
torchaudio>=2.0.1
torchmetrics>=0.7.2
torchsampler>=0.1.1
torchvision>=0.11.2
tornado==6.1
tqdm==4.63.0
traitlets==5.9.0
triton==2.0.0
ttach==0.0.3
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
unicodedata2==14.0.0
uri-template==1.2.0
urllib3==1.26.8
validators==0.18.2
wandb==0.12.7
watchdog==2.1.7
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
widgetsnbextension==3.6.0
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
